ANN-1 create the project folder structure
Create the project architecture according to the architecture created by
Ömer.

This structure contains the activation folder for the activation
functions and the variants of the activation functions, each with a .py
file in it. It also contains the loss folder for the loss functions, the
optimizer folder for the optimizer functions, and the summation folder
for the summation functions. The internal structure of these folders is
exactly the same as the activation folder. It also contains the layer
folder and the Dense file to create the layers, the model folder to
create the main model and the Network file inside it. Finally, there is
the test.py file in the model folder to run the tests.

Closes ANN-1