feat(init): create uniform weight init. technique
Create uniform weight initialization technique. With this commit, create
the starting values of the required weights for each layer created in
the artificial neural network structure according to the uniform method.

This commit;
- Replaces the weight class in the init folder in the main project
structure with uniform_weight,
- creates initializer.py file,
- Creates the Uniform class in the uniform_weight.py file and the
__init__() and build() methods in that class. When the build() method
is triggered, it returns the initial values of the weight in the number
of size arguments it receives, in accordance with the uniform technique,
- Creates Initializer class in initializer.py file. This class is a
class in the frozen dictionary structure. The object it consists of
cannot be changed in any way after it is first defined,
- Inside the initializer.py file, it also creates an object named
initializer from the Initializer class and contains the uniform key and
the Uniform() value.