feat(layer): create Dense class and first __init__ features
Create Dense class with __init__ method and this first features.

The __init__ method takes 5 arguments. The neuron argument with the
default value of 4, the weight_init argument with the uniform default
literal, the bias_init argument with the zeros default literal, the
summation argument with the sum default literal, and the activation
argument with the relu default literal. It also assigns weight_init and
bias_init parameters to class-level variables with the same name.

Refs ANN-3