feat(init): create zero bias init. technique
Create zero bias initialization technique. With this commit, create the
starting values of the required bias for each layer created in
the artificial neural network structure according to the zero method.

This commit;
- Replaces the bias class in the init folder in the main project
structure with zero_bias,
- Creates the Zero class in the zero_bias.py file and the
__init__() and build() methods in that class. When the build() method
is triggered, it returns the initial value of the bias in accordance
with the zero technique,
- Inside the initializer.py file, the initalizer defines the key zero
and the value Zero() to be assigned to the frozen dictionary object at
startup.

Refs ANN-5