feat(summation): create sum summation init. technique
Create sum summation initialization technique

This commit;
- Creates initializer.py file in summation folder,
- Creates the Sum class in the sum.py file and the
__init__() and build() methods in that class. When the build() method
is triggered, it returns the summation value of the weight*inputs lists
in arguments it receives, in accordance with the sum technique,
- Creates Initializer class in initializer.py file. This class is a
class in the frozen dictionary structure. The object it consists of
cannot be changed in any way after it is first defined,
- Inside the initializer.py file, it also creates an object named
summation from the Initializer class and contains the sum key and
the Sum() value.

Refs ANN-6