feat(activation): create ReLu activation function
Create ReLu activation function. With this commit, apply the activation
function according to the ReLu mathematical expression to the output of
each neuron in the relevant layer in the neural network.

This commit;
- Creates initializer.py file in activation folder,
- Creates the ReLu class in the relu.py file and the
__init__() and build() methods in that class. When the build() method
is triggered, it returns the activation value of the neuron output
in arguments it receives, in accordance with the relu method,
- Creates Initializer class in initializer.py file. This class is a
class in the frozen dictionary structure. The object it consists of
cannot be changed in any way after it is first defined,
- Inside the initializer.py file, it also creates an object named
activation from the Initializer class and contains the relu key and
the ReLu() value.

Refs ANN-7