feat(model): add build() class method
Add build() class method in Network class. Also the build() method
tested in test.py file.

In the build() method, it is first checked whether the network has been
created and whether the network contains at least 1 hidden layer. If
these are OK, then the hyperparameters of the input layer are reset
because they are not needed. As the last and main code, the weight and
bias values of the hidden layer and output layer are initialized and the
network is also initialized.