feat(model): add forward() method in Network class
Add forward() method in Network class. This method allows the forward
feed of the network. First, it calls the build() method if the network
is not built. It then completes the forward feeding of the network by
calling and implementing all necessary network hyperparameters in order
according to the size of the network. It returns a one-dimensional
result array the size of the last layer's neuron. Also test this method
in test.py.

Refs ANN-8