feat(loss): create MSE loss function
Create MSE loss function. With this commit, apply the loss
function according to the MSE mathematical expression to the output of
output layer in the neural network.

This commit;
- Creates initializer.py file in loss folder,
- Creates the MeanSquaredError class in the mean_squared_error.py file
and the __init__() and calculate() methods in that class. When the
calculate() method is triggered, it returns the loss value of the
output layer in arguments it receives which is yhat, in accordance with
the mse method,
- Creates Initializer class in initializer.py file. This class is a
class in the frozen dictionary structure. The object it consists of
cannot be changed in any way after it is first defined,
- Inside the initializer.py file, it also creates an object named
loss from the Initializer class and contains the mse key and
the MeanSquaredError() value.

Refs ANN-9