feat(loss): edit calculate() method in MeanSquaredError class
Edit calculate() method in MeanSquaredError class. The use of if and
elif in this method becomes unnecessary because the numpy functions in
elif condition give correct and seamless results in both cases whether
the int value comes from the arguments it receives or nump.ndarray.
This commit completely removes the if construct.